# simple optimizations