- Copy prebuilt default flags
- Use dataclasses with slots instead of nametuples

Before:
multithreaded: Overall: 124463.31 RPS / 8.03 us/req
singlethreaded: Overall: 126979.22 RPS / 7.88 us/req

After:
multithreaded: Overall: 135509.96 RPS / 7.38 us/req
singlethreaded: Overall: 137630.69 RPS / 7.27 us/req